Changed settings to reload game when modified
The system settings drive changes to the document sheets, and when they
are changed, the documents must be closed and opened again to display
the changes. This behavior might cause confusion, so now every setting
that pertains to changes to the document sheets cause the game to reload
for all clients.